requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.31.0
qrcode==7.4.2
Pillow==10.1.0
dj-database-url==2.1.0

# Django and related packages
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
django-environ==0.11.2
drf-yasg==1.21.10
django-prometheus==2.3.1

# Testing
coverage==7.4.1
